Add configurable node drain timeouts
Defaults to 0 which means it will wait indefinitely. Value is the
integer number of seconds to wait before moving on to the next step in
the playbook. When the timeout is hit the return code will be non zero
and will include a message indicating that the timeout has been
exceeded. If the output does not contain that message there's likely
some other problem so treat that situation as a failure.